+++ extensions/crossmodel-lang/src/main.ts
@@ -3,12 +3,10 @@
  ********************************************************************************/
 import 'reflect-metadata';
 
+import { createCrossModelServices, startGLSPServer, startModelServer } from '@crossmodel/server';
 import { startLanguageServer } from 'langium/lsp';
 import { NodeFileSystem } from 'langium/node';
 import { createConnection, ProposedFeatures } from 'vscode-languageserver/node.js';
-import { startGLSPServer } from './glsp-server/launch.js';
-import { createCrossModelServices } from './language-server/cross-model-module.js';
-import { startModelServer } from './model-server/launch.js';
 
 /**
  * This module will be spawned as a separate language server process by the 'extension.ts'.