+++ packages/server/src/glsp-server/mapping-diagram/command-palette/add-source-object-action-provider.ts
@@ -1,0 +1,26 @@
+/********************************************************************************
+ * Copyright (c) 2023 CrossBreeze.
+ ********************************************************************************/
+import { AddSourceObjectOperation, codiconCSSString } from '@crossmodel/protocol';
+import { Args, CommandPaletteActionProvider, GModelElement, LabeledAction, Point } from '@eclipse-glsp/server';
+import { injectable } from 'inversify';
+import { SourceObject } from '../../../language-server/generated/ast.js';
+import { MappingModelState } from '../model/mapping-model-state.js';
+
+@injectable()
+export class MappingDiagramCommandPaletteActionProvider extends CommandPaletteActionProvider {
+   declare protected modelState: MappingModelState;
+
+   getPaletteActions(_selectedElementIds: string[], _selectedElements: GModelElement[], position: Point, args?: Args): LabeledAction[] {
+      const completionItems = this.modelState.services.language.references.ScopeProvider.complete({
+         container: { globalId: this.modelState.mapping.id! },
+         syntheticElements: [{ property: 'sources', type: SourceObject }],
+         property: 'entity'
+      });
+      return completionItems.map<LabeledAction>(item => ({
+         label: item.label,
+         actions: [AddSourceObjectOperation.create(item.label, position || Point.ORIGIN)],
+         icon: codiconCSSString('inspect')
+      }));
+   }
+}