+++ packages/core/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.5](https://github.com/crossmodel/crossmodel-core/compare/v0.7.4...v0.7.5) (2025-09-04)
+
+**Note:** Version bump only for package @crossmodel/core
+
+
+
+
+
+## [0.7.4](https://github.com/crossmodel/crossmodel-core/compare/v0.7.3...v0.7.4) (2025-08-29)
+
+**Note:** Version bump only for package @crossmodel/core
+
+
+
+
+
+## [0.7.2](https://github.com/crossmodel/crossmodel-core/compare/v0.7.1...v0.7.2) (2025-08-21)
+
+**Note:** Version bump only for package @crossmodel/core
+
+
+
+
+
 ## [0.7.1](https://github.com/crossmodel/crossmodel-core/compare/v0.7.0...v0.7.1) (2025-08-05)
 
 **Note:** Version bump only for package @crossmodel/core